admin_panel/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.4
Werkzeug==3.0.1